backend/requirements.txt
Flask==2.0.1
flask_sqlalchemy==2.5.1
psycopg2-binary==2.9.1
geoalchemy2==0.8.4
python-dotenv==0.19.2
